chore: Updated example app -  jetpack compose
