+++ examples/basic-integration-jetpack-compose/app/build.gradle
@@ -60,6 +60,7 @@ dependencies {
 
     // DEUNA SDK
     implementation 'com.github.deuna-developers:deuna-sdk-android:2.7.0'
+
     // Material Design 3
     implementation 'androidx.compose.material3:material3'
