+++ examples/basic-integration-jetpack-compose/app/src/main/java/com/deuna/compose_demo/MainActivity.kt
@@ -5,7 +5,7 @@ import androidx.activity.*
 import androidx.activity.compose.*
 import androidx.navigation.compose.*
 import com.deuna.compose_demo.screens.*
-import com.deuna.compose_demo.view_models.*
+import com.deuna.compose_demo.view_models.home.HomeViewModel
 import com.deuna.maven.*
 import com.deuna.maven.shared.Environment
 import com.deuna.maven.shared.toMap
