--- examples/basic-integration-jetpack-compose/app/src/main/java/com/deuna/compose_demo/view_models/HomeViewModel.kt
@@ -1,289 +1,0 @@
-package com.deuna.compose_demo.view_models
-
-import android.content.Context
-import android.util.Log
-import androidx.compose.runtime.*
-import androidx.lifecycle.*
-import com.deuna.compose_demo.screens.CheckoutResult
-import com.deuna.compose_demo.screens.ElementsResult
-import com.deuna.compose_demo.screens.PaymentWidgetResult
-import com.deuna.maven.*
-import com.deuna.maven.checkout.domain.*
-import com.deuna.maven.element.domain.ElementsError
-import com.deuna.maven.payment_widget.domain.PaymentWidgetCallbacks
-import com.deuna.maven.shared.*
-import com.deuna.maven.shared.domain.UserInfo
-import kotlinx.coroutines.*
-import org.json.JSONObject
-
-const val ERROR_TAG = "❌ DeunaSDK"
-const val DEBUG_TAG = "👀 DeunaSDK"
-
-/**
- * ViewModel for the Home screen, responsible for handling user interactions and data manipulation.
- * @param deunaSDK The DeunaSDK instance used for payment and card saving operations.
- */
-@Suppress("UNCHECKED_CAST")
-class HomeViewModel(private val deunaSDK: DeunaSDK) : ViewModel() {
-
-    // State variables for order token and user token
-    val orderToken = mutableStateOf("")
-    val userToken = mutableStateOf("")
-
-    val userTokenValue: String?
-        get() {
-            return userToken.value.ifEmpty { null }
-        }
-
-
-    fun showPaymentWidget(
-        context: Context,
-        completion: (PaymentWidgetResult) -> Unit,
-    ) {
-        deunaSDK.initPaymentWidget(
-            context = context,
-            orderToken = orderToken.value.trim(),
-            callbacks = paymentWidgetsCallbacks(completion),
-            userToken = userTokenValue,
-            cssFile = "YOUR_THEME_UUID", // optional
-            paymentMethods = listOf(
-                mapOf(
-                    "payment_method" to "voucher",
-                    "processors" to listOf("daviplata", "nequi_push_voucher")
-                ), mapOf(
-                    "payment_method" to "paypal", "processors" to listOf("paypal")
-                )
-            )
-        )
-    }
-
-
-    private fun checkoutCallbacks(
-        completion: (CheckoutResult) -> Unit,
-    ): CheckoutCallbacks {
-        return CheckoutCallbacks().apply {
-            onSuccess = { data ->
-                deunaSDK.close()
-                viewModelScope.launch {
-                    completion(
-                        CheckoutResult.Success(
-                            data["order"] as Json
-                        )
-                    )
-                }
-            }
-            onError = { error ->
-                Log.e(ERROR_TAG, "on error ${error.type} , ${error.metadata}")
-                when (error.type) {
-                    PaymentsError.Type.PAYMENT_ERROR, PaymentsError.Type.ORDER_COULD_NOT_BE_RETRIEVED, PaymentsError.Type.NO_INTERNET_CONNECTION -> {
-                        deunaSDK.close()
-                        completion(CheckoutResult.Error(error))
-                    }
-
-                    else -> {}
-                }
-            }
-            onCanceled = {
-                viewModelScope.launch {
-                    completion(CheckoutResult.Canceled)
-                }
-            }
-            onEventDispatch = { event, data ->
-                Log.d(DEBUG_TAG, "onEventDispatch ${event.name}: $data")
-                when (event) {
-                    CheckoutEvent.changeCart, CheckoutEvent.changeAddress -> {
-                        deunaSDK.close()
-                        viewModelScope.launch {
-                            completion(CheckoutResult.Canceled)
-                        }
-                    }
-
-                    else -> Log.d(DEBUG_TAG, "on event ${event.value}")
-                }
-            }
-        }
-    }
-
-    /**
-     * Initiates the payment process.
-     * @param context The activity context.
-     * @param completion Callback to be invoked upon completion of the payment process.
-     */
-    fun showCheckout(
-        context: Context,
-        completion: (CheckoutResult) -> Unit,
-    ) {
-        deunaSDK.initCheckout(
-            context = context,
-            orderToken = orderToken.value.trim(),
-            callbacks = checkoutCallbacks(completion),
-            userToken = userTokenValue,
-        )
-    }
-
-
-    private fun paymentWidgetsCallbacks(
-        completion: (PaymentWidgetResult) -> Unit,
-    ): PaymentWidgetCallbacks {
-        return PaymentWidgetCallbacks().apply {
-            onSuccess = { data ->
-                deunaSDK.close()
-                viewModelScope.launch {
-                    completion(
-                        PaymentWidgetResult.Success(
-                            data["order"] as Json
-                        )
-                    )
-                }
-            }
-            onCanceled = {
-                viewModelScope.launch {
-                    completion(PaymentWidgetResult.Canceled)
-                }
-            }
-            onError = { error ->
-                Log.e(ERROR_TAG, "on error ${error.type} , ${error.metadata}")
-                when (error.type) {
-                    PaymentsError.Type.INITIALIZATION_FAILED, PaymentsError.Type.NO_INTERNET_CONNECTION -> {
-                        deunaSDK.close()
-                        completion(PaymentWidgetResult.Error(error))
-                    }
-
-                    else -> {}
-                }
-
-            }
-            onClosed = {
-
-            }
-            onCardBinDetected = { cardBinMetadata, refetchOrder ->
-                deunaSDK.setCustomStyle(
-                    data = JSONObject(
-                        """
-                        {
-                          "theme": {
-                            "colors": {
-                              "primaryTextColor": "#023047",
-                              "backgroundSecondary": "#8ECAE6",
-                              "backgroundPrimary": "#F2F2F2",
-                              "buttonPrimaryFill": "#FFB703",
-                              "buttonPrimaryHover": "#FFB703",
-                              "buttonPrimaryText": "#000000",
-                              "buttonPrimaryActive": "#FFB703"
-                            }
-                          },
-                          "HeaderPattern": {
-                            "overrides": {
-                              "Logo": {
-                                "props": {
-                                  "url": "https://images-staging.getduna.com/ema/fc78ef09-ffc7-4d04-aec3-4c2a2023b336/test2.png"
-                                }
-                              }
-                            }
-                          }
-                        }
-                        """
-                    ).toMap()
-                )
-            }
-            onPaymentProcessing = {
-                Log.d(DEBUG_TAG, "onPaymentProcessing")
-            }
-            onEventDispatch = { event, data ->
-                Log.d(DEBUG_TAG, "onEventDispatch ${event.name}: $data")
-            }
-        }
-    }
-
-
-    /**
-     * Show the widget that processes a payment using Click to Pay.
-     */
-    fun clickToPay(
-        context: Context,
-        completion: (ElementsResult) -> Unit,
-    ) {
-        deunaSDK.initElements(
-            context = context,
-            userInfo = UserInfo(
-                firstName = "Darwin", lastName = "Morocho", email = "dmorocho+10@deuna.com"
-            ),
-            types = listOf(
-                mapOf(
-                    "name" to ElementsWidget.CLICK_TO_PAY
-                )
-            ),
-            callbacks = ElementsCallbacks().apply {
-                onSuccess = {
-                    deunaSDK.close()
-                    viewModelScope.launch {
-                        completion(ElementsResult.Success(it["metadata"] as Json))
-                    }
-                }
-                onError = {
-                    if (it.type == ElementsError.Type.INITIALIZATION_FAILED) {
-                        deunaSDK.close()
-                        viewModelScope.launch {
-                            completion(ElementsResult.Error(it))
-                        }
-                    }
-                }
-                onEventDispatch = { event, data ->
-                    Log.d(DEBUG_TAG, "onEventDispatch ${event.name}: $data")
-                }
-            },
-        )
-    }
-
-    /**
-     * Initiates the process of saving the card information.
-     * @param context The activity context.
-     * @param completion Callback to be invoked upon completion of the card saving process.
-     */
-    fun saveCard(
-        context: Context,
-        completion: (ElementsResult) -> Unit,
-    ) {
-        deunaSDK.initElements(
-            context = context,
-            userToken = userTokenValue,
-            userInfo = if (userTokenValue == null) UserInfo(
-                firstName = "Darwin", lastName = "Morocho", email = "dmorocho@deuna.com"
-            ) else null,
-            callbacks = elementsCallbacks(completion)
-        )
-    }
-
-
-    private fun elementsCallbacks(
-        completion: (ElementsResult) -> Unit,
-    ): ElementsCallbacks {
-        return ElementsCallbacks().apply {
-            onSuccess = { response ->
-                deunaSDK.close()
-                viewModelScope.launch {
-                    completion(
-                        ElementsResult.Success(
-                            (response["metadata"] as Json)["createdCard"] as Json
-                        )
-                    )
-                }
-            }
-            onError = { error ->
-                deunaSDK.close()
-                viewModelScope.launch {
-                    completion(ElementsResult.Error(error))
-                }
-            }
-            onCanceled = {
-                viewModelScope.launch {
-                    completion(ElementsResult.Canceled)
-                }
-            }
-            onEventDispatch = { event, data ->
-                Log.d(DEBUG_TAG, "onEventDispatch ${event.name}: $data")
-            }
-        }
-    }
-}
-
