+++ examples/basic-integration-jetpack-compose/app/src/main/java/com/deuna/compose_demo/view_models/home/initElementsVault.kt
@@ -1,0 +1,57 @@
+package com.deuna.compose_demo.view_models.home
+
+import android.content.Context
+import android.util.Log
+import androidx.lifecycle.viewModelScope
+import com.deuna.compose_demo.screens.ElementsResult
+import com.deuna.maven.initElements
+import com.deuna.maven.shared.ElementsCallbacks
+import com.deuna.maven.shared.Json
+import com.deuna.maven.shared.domain.UserInfo
+import kotlinx.coroutines.launch
+
+/**
+ * Initiates the process of saving the card information.
+ * @param context The activity context.
+ * @param completion Callback to be invoked upon completion of the card saving process.
+ */
+fun HomeViewModel.saveCard(
+    context: Context,
+    completion: (ElementsResult) -> Unit,
+) {
+    deunaSDK.initElements(
+        context = context,
+        userToken = userTokenValue,
+        userInfo = if (userTokenValue == null) UserInfo(
+            firstName = "Darwin", lastName = "Morocho", email = "dmorocho@deuna.com"
+        ) else null,
+        callbacks = ElementsCallbacks().apply {
+            onSuccess = { response ->
+                deunaSDK.close()
+                viewModelScope.launch {
+                    completion(
+                        ElementsResult.Success(
+                            (response["metadata"] as Json)["createdCard"] as Json
+                        )
+                    )
+                }
+            }
+            onError = { error ->
+                deunaSDK.close()
+                viewModelScope.launch {
+                    completion(ElementsResult.Error(error))
+                }
+            }
+            onCanceled = {
+                viewModelScope.launch {
+                    completion(ElementsResult.Canceled)
+                }
+            }
+            onEventDispatch = { event, data ->
+                Log.d(DEBUG_TAG, "onEventDispatch ${event.name}: $data")
+            }
+        }
+    )
+}
+
+
